+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 * Select by primary index name
+* Fix error handling select with invalid type value 
 
 ## [0.2.0] - 2020-10-07